+++ packages/blade/src/utils/storybook/Sandbox/types.ts
@@ -1,4 +1,21 @@
 import type { BaseBoxProps } from '~components/Box/BaseBox';
+import type { Platform } from '~utils/platform';
+
+type SandboxStackBlitzProps = {
+  children: string;
+  editorHeight?: number | string;
+  padding?: BaseBoxProps['padding'];
+  showConsole?: boolean;
+  /**
+   * Renders link in react native stories to open storybook on web
+   *
+   * (Its workaround since sandpack doesn't work on native)
+   */
+  uri?: Platform.Select<{
+    web: never;
+    native: string;
+  }>;
+};
 
 type SandboxProps = {
   children: string;
@@ -33,4 +50,4 @@ type RecipeSandboxProps = {
   view?: 'preview' | 'editor';
 };
 
-export type { SandboxProps, RecipeSandboxProps };
+export type { SandboxProps, SandboxStackBlitzProps, RecipeSandboxProps };