+++ packages/blade/src/utils/storybook/StoryPageWrapper.tsx
@@ -3,7 +3,7 @@ import { ArgsTable, Primary, PRIMARY_STORY, Stories } from '@storybook/addon-doc
 import styled from 'styled-components';
 import useMakeFigmaURL from './useMakeFigmaURL';
 import FigmaEmbed from './FigmaEmbed';
-import { SandboxHighlighter } from './Sandbox';
+import { SandboxHighlighter } from './Sandbox/SandpackEditor';
 import { componentData } from './componentStatusData';
 import BaseBox from '~components/Box/BaseBox';
 import { Alert } from '~components/Alert';
@@ -25,12 +25,14 @@ type StoryPageWrapperTypes = {
     paymentTheme: string;
     bankingTheme: string;
   };
+  argTableComponent?: unknown;
   componentDescription: string;
   propsDescription?: string;
   componentName: string;
   children?: React.ReactNode;
   note?: React.ReactChild;
   showStorybookControls?: boolean;
+  showDefaultExample?: boolean;
   showArgsTable?: boolean;
   /**
    * Use this to override default API decision link generated from componentName
@@ -40,6 +42,10 @@ type StoryPageWrapperTypes = {
    * Use this to override default imports
    */
   imports?: string;
+  /**
+   * Use this to override default API decision component name
+   */
+  apiDecisionComponentName?: string;
 };
 
 // Global Styles are not applied by default on `.mdx` pages of storybook. So just overriding few global styles here which are applied to both, stories and guide pages
@@ -92,7 +98,7 @@ const StoryPageWrapper = (props: StoryPageWrapperTypes): React.ReactElement => {
     (componentInfo) => componentInfo.name === props.componentName,
   );
 
-  const { showStorybookControls = true, showArgsTable = true } = props;
+  const { showStorybookControls = true, showArgsTable = true, showDefaultExample = true } = props;
 
   return (
     <BladeProvider themeTokens={paymentTheme}>
@@ -145,11 +151,15 @@ const StoryPageWrapper = (props: StoryPageWrapperTypes): React.ReactElement => {
         )}
         {showStorybookControls ? (
           <>
-            <Title size="large">Example</Title>
-            <Subtitle size="medium" marginY="spacing.4">
-              {`This is the default ${props.componentName}. You can change the properties using the controls below.`}
-            </Subtitle>
-            <Primary />
+            {showDefaultExample ? (
+              <>
+                <Title size="large">Example</Title>
+                <Subtitle size="medium" marginY="spacing.4">
+                  {`This is the default ${props.componentName}. You can change the properties using the controls below.`}
+                </Subtitle>
+                <Primary />
+              </>
+            ) : null}
             {showArgsTable ? (
               <>
                 <BaseBox id="properties-ref">
@@ -164,7 +174,7 @@ const StoryPageWrapper = (props: StoryPageWrapperTypes): React.ReactElement => {
                           `https://github.com/razorpay/blade/blob/master/packages/blade/src/components/${props.componentName}/_decisions/decisions.md`
                         }
                       >
-                        API Decisions for {props.componentName}
+                        API Decisions for {props.apiDecisionComponentName ?? props.componentName}
                       </Link>
                     </Text>
                   )}
@@ -180,7 +190,10 @@ const StoryPageWrapper = (props: StoryPageWrapperTypes): React.ReactElement => {
                     </Box>
                   ) : null}
                 </BaseBox>
-                <ArgsTable story={PRIMARY_STORY} />
+                <ArgsTable
+                  story={props.argTableComponent ? undefined : PRIMARY_STORY}
+                  of={props.argTableComponent}
+                />
               </>
             ) : null}