+++ packages/blade/src/utils/useScrollLock.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
+import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock-upgrade';
 import React from 'react';
 
 type UseScrollLockProps = {
-  targetRef: React.RefObject<Element>;
+  targetRef?: React.RefObject<Element>;
   enabled: boolean;
   reserveScrollBarGap: boolean;
 };
@@ -53,7 +53,7 @@ export function useScrollLock({
       return;
     }
 
-    const target = targetRef.current;
+    const target = (targetRef ? targetRef.current : document.body) as HTMLElement;
     active.current = false;
 
     ref.current = {